# apps/web/public/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Ledger Agent Payments API",
		"version": "0.2.0",
		"description": "Agent-first API for submitting transaction intents. AI agents propose transactions; humans review and sign on Ledger hardware. Agents never touch private keys.\n\n## Quick Start\n1. Register your agent key (human signs authorization on Ledger)\n2. `POST /api/intents` with AgentAuth header to propose a transaction\n3. Share the returned `paymentUrl` with the human\n4. Poll `GET /api/intents/{id}` until status is `confirmed`, `rejected`, `failed`, or `expired`\n\n## Authentication\nAgents authenticate via the `Authorization: AgentAuth <timestamp>.<bodyHash>.<signature>` header.\n- `timestamp`: Unix epoch seconds (within 5 min of server time)\n- `bodyHash`: keccak256 of raw JSON body as hex (use `0x` for GET)\n- `signature`: EIP-191 personal_sign of `<timestamp>.<bodyHash>` with agent's secp256k1 key"
	},
	"servers": [
		{ "url": "https://agent-intents-web.vercel.app", "description": "Production" },
		{ "url": "/", "description": "Same-origin" }
	],
...
	"paths": {
...
		"/api/intents": {
			"get": {
				"tags": ["Intents"],
				"operationId": "listIntents",
				"summary": "List intents for a user",
				"description": "Returns intents for a given userId, optionally filtered by status. No authentication required.",
				"parameters": [
					{
						"name": "userId",
						"in": "query",
						"required": true,
						"description": "Wallet address (e.g., 0xabc...)",
						"schema": { "type": "string" }
					},
					{
						"name": "status",
						"in": "query",
						"required": false,
						"description": "Filter by intent status",
						"schema": { "$ref": "#/components/schemas/IntentStatus" }
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"description": "Max results (1–100)",
						"schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/IntentListResponse" }
							}
						}
					},
					"400": {
						"description": "Missing userId",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				}
			},
			"post": {
				"tags": ["Intents"],
				"operationId": "createIntent",
				"summary": "Create a payment intent",
				"description": "Submit a new transaction intent for human review and Ledger signing. Requires AgentAuth. The userId is derived from the authenticated agent's trustchain — do not pass it in the body.\n\nReturns a `paymentUrl` that the agent should share with the human so they can review and sign.",
				"security": [{ "AgentAuth": [] }],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/CreateIntentRequest" },
							"examples": {
								"simple_transfer": {
									"summary": "Simple USDC transfer",
									"value": {
										"agentId": "my-research-bot",
										"agentName": "Research Assistant",
										"urgency": "normal",
										"expiresInMinutes": 60,
										"details": {
											"type": "transfer",
											"token": "USDC",
											"amount": "5.00",
											"recipient": "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045",
											"chainId": 8453,
											"memo": "API access fee"
										}
									}
								},
								"x402_payment": {
									"summary": "x402 pay-per-call API payment",
									"value": {
										"agentId": "my-research-bot",
										"agentName": "Research Assistant",
										"details": {
											"type": "transfer",
											"token": "USDC",
											"amount": "0.01",
											"recipient": "0xPayToAddress...",
											"chainId": 8453,
											"memo": "x402 payment for CoinGecko Pro API",
											"resource": "https://api.coingecko.com/pro/v1/coins/bitcoin",
											"category": "api_payment",
											"x402": {
												"resource": { "url": "https://api.coingecko.com/pro/v1/coins/bitcoin" },
												"accepted": {
													"scheme": "exact",
													"network": "eip155:8453",
													"amount": "10000",
													"asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
													"payTo": "0xPayToAddress..."
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Intent created",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/CreateIntentResponse" }
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"401": {
						"description": "Missing or invalid AgentAuth header",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"429": {
						"description": "Rate limit exceeded (max 10 intents per agent per minute)",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"503": {
						"description": "Service temporarily unavailable",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				}
			}
		},
...
		"/api/intents/status": {
			"post": {
				"tags": ["Intents"],
				"operationId": "updateIntentStatus",
				"summary": "Update intent status",
				"description": "Update the status of an intent. Agents (AgentAuth) can set: `executing`, `confirmed`, `failed`. Humans (session cookie) can set: `approved`, `rejected`, `authorized`, `broadcasting`. The caller must own the intent.\n\nThis is the preferred route (static path avoids Vercel dynamic-route issues).",
				"security": [{ "AgentAuth": [] }, { "SessionCookie": [] }],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/UpdateIntentStatusByIdRequest" },
							"examples": {
								"agent_confirms": {
									"summary": "Agent confirms x402 settlement",
									"value": {
										"id": "int_1707048000_abc12345",
										"status": "confirmed",
										"note": "x402 payment settled successfully",
										"settlementReceipt": {
											"txHash": "0xabc123...",
											"network": "eip155:8453",
											"success": true
										}
									}
								},
								"agent_reports_failure": {
									"summary": "Agent reports failure",
									"value": {
										"id": "int_1707048000_abc12345",
										"status": "failed",
										"note": "x402 server returned 500 after payment"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Status updated",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/IntentResponse" } }
						}
					},
					"400": {
						"description": "Invalid status or missing fields",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"401": {
						"description": "Authentication failed",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Not authorized (wrong owner or disallowed status transition)",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "Intent not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"409": {
						"description": "Invalid state transition (e.g., confirmed → pending)",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				}
			}
		},
...
		"/api/agents/register": {
			"post": {
				"tags": ["Agents"],
				"operationId": "registerAgent",
				"summary": "Register a new agent",
				"description": "Register a new agent public key under a trustchain identity. No auth header needed — authorization is proved via an EIP-191 signature in the body (signed on the Ledger device).\n\nThe signed message must be exactly:\n```\nAuthorize agent key for Ledger Agent Payments\nKey: <agentPublicKey>\nLabel: <agentLabel>\nIdentity: <trustChainId>\n```",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/RegisterAgentRequest" },
							"examples": {
								"register": {
									"summary": "Register a trading bot",
									"value": {
										"trustChainId": "0xabcdef1234567890abcdef1234567890abcdef12",
										"agentPublicKey": "0x02abc123def456789...",
										"agentLabel": "My Trading Bot",
										"authorizationSignature": "0xsig..."
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Agent registered",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/AgentResponse" } }
						}
					},
					"400": {
						"description": "Validation error or invalid signature",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Signature does not match wallet",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"409": {
						"description": "Agent public key already registered",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"429": {
						"description": "Rate limit exceeded (max 5 registrations per wallet per minute)",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"503": {
						"description": "Service temporarily unavailable",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"AgentAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization",
				"description": "Agent authentication. Format: `AgentAuth <timestamp>.<bodyHash>.<signature>`\n\n- `timestamp`: Unix epoch seconds (must be within 5 min of server time)\n- `bodyHash`: keccak256 of the raw JSON body as hex string (use `0x` for GET requests)\n- `signature`: EIP-191 personal_sign of the string `<timestamp>.<bodyHash>` using the agent's secp256k1 private key\n\nThe server recovers the signer address and matches it against registered agent public keys."
			},
			"SessionCookie": {
				"type": "apiKey",
				"in": "cookie",
				"name": "ai_session",
				"description": "Session cookie established via POST /api/auth/verify. Valid for 7 days. Used by the web UI for human actions."
			}
		},
		"schemas": {
...
			"ErrorResponse": {
				"type": "object",
				"required": ["success", "error"],
				"properties": {
					"success": { "type": "boolean", "const": false },
					"error": { "type": "string", "description": "Human-readable error message" }
				}
			},
			"IntentUrgency": {
				"type": "string",
				"enum": ["low", "normal", "high", "critical"],
				"default": "normal",
				"description": "Priority level. Affects recommended polling interval."
			},
			"IntentStatus": {
				"type": "string",
				"enum": [
					"pending",
					"approved",
					"rejected",
					"broadcasting",
					"authorized",
					"executing",
					"confirmed",
					"failed",
					"expired"
				],
				"description": "Intent lifecycle status. Terminal states: rejected, confirmed, failed, expired."
			},
			"PaymentCategory": {
				"type": "string",
				"enum": [
					"api_payment",
					"subscription",
					"purchase",
					"p2p_transfer",
					"defi",
					"bill_payment",
					"donation",
					"other"
				],
				"description": "Category of payment (for display and analytics)"
			},
...
			"TransferIntent": {
				"type": "object",
				"required": ["type", "token", "amount", "recipient", "chainId"],
				"properties": {
					"type": { "type": "string", "const": "transfer" },
					"token": { "type": "string", "description": "Token symbol (e.g., USDC, ETH)" },
					"tokenAddress": {
						"type": "string",
						"description": "ERC-20 contract address (auto-filled for known tokens)"
					},
					"tokenLogo": { "type": "string", "description": "URL to token logo image" },
					"amount": {
						"type": "string",
						"description": "Human-readable amount (e.g., '50', '0.01')"
					},
					"amountWei": { "type": "string", "description": "Amount in wei for precision" },
					"recipient": { "type": "string", "description": "Destination address (0x...)" },
					"recipientEns": { "type": "string", "description": "ENS name if resolved" },
					"chainId": {
						"type": "integer",
						"description": "Chain ID: 8453 (Base), 11155111 (Sepolia), 84532 (Base Sepolia)"
					},
					"memo": { "type": "string", "description": "Human-readable reason for the transaction" },
					"resource": {
						"type": "string",
						"description": "x402: URL of the resource being paid for"
					},
					"merchant": {
						"type": "object",
						"description": "Merchant/payee information",
						"properties": {
							"name": { "type": "string" },
							"url": { "type": "string" },
							"logo": { "type": "string" },
							"verified": { "type": "boolean" }
						}
					},
					"category": { "$ref": "#/components/schemas/PaymentCategory" },
					"x402": { "$ref": "#/components/schemas/X402Context" }
				}
			},
			"Intent": {
				"type": "object",
				"required": [
					"id",
					"userId",
					"agentId",
					"agentName",
					"details",
					"urgency",
					"status",
					"createdAt",
					"statusHistory"
				],
				"properties": {
					"id": {
						"type": "string",
						"description": "Unique intent ID (format: int_<timestamp>_<uuid>)"
					},
					"userId": { "type": "string", "description": "Wallet address of the human signer" },
					"agentId": { "type": "string", "description": "Agent identifier" },
					"agentName": { "type": "string", "description": "Agent display name" },
					"details": { "$ref": "#/components/schemas/TransferIntent" },
					"urgency": { "$ref": "#/components/schemas/IntentUrgency" },
					"status": { "$ref": "#/components/schemas/IntentStatus" },
					"trustChainId": {
						"type": "string",
						"description": "Trustchain identity (wallet address) — set when created by an authenticated agent"
					},
					"createdByMemberId": {
						"type": "string",
						"format": "uuid",
						"description": "Agent member UUID"
					},
					"createdAt": { "type": "string", "format": "date-time" },
					"expiresAt": { "type": "string", "format": "date-time" },
					"reviewedAt": { "type": "string", "format": "date-time" },
					"broadcastAt": { "type": "string", "format": "date-time" },
					"confirmedAt": { "type": "string", "format": "date-time" },
					"txHash": {
						"type": "string",
						"description": "Transaction hash (populated after confirmation)"
					},
					"txUrl": { "type": "string", "description": "Block explorer URL" },
					"statusHistory": {
						"type": "array",
						"description": "Ordered audit trail of status changes",
						"items": {
							"type": "object",
							"required": ["status", "timestamp"],
							"properties": {
								"status": { "$ref": "#/components/schemas/IntentStatus" },
								"timestamp": { "type": "string", "format": "date-time" },
								"note": { "type": "string" }
							}
						}
					}
				}
			},
			"IntentResponse": {
				"type": "object",
				"required": ["success", "intent"],
				"properties": {
					"success": { "type": "boolean", "const": true },
					"intent": { "$ref": "#/components/schemas/Intent" }
				}
			},
			"CreateIntentResponse": {
				"type": "object",
				"required": ["success", "intent", "paymentUrl"],
				"properties": {
					"success": { "type": "boolean", "const": true },
					"intent": { "$ref": "#/components/schemas/Intent" },
					"paymentUrl": {
						"type": "string",
						"format": "uri",
						"description": "Direct URL for the human to review and sign this intent. Share this with the user."
					}
				}
			},
			"IntentListResponse": {
				"type": "object",
				"required": ["success", "intents"],
				"properties": {
					"success": { "type": "boolean", "const": true },
					"intents": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/Intent" }
					}
				}
			},
			"CreateIntentRequest": {
				"type": "object",
				"required": ["agentId", "details"],
				"properties": {
					"agentId": { "type": "string", "description": "Your agent's unique identifier" },
					"agentName": {
						"type": "string",
						"description": "Human-readable display name (defaults to agentId)"
					},
					"urgency": { "$ref": "#/components/schemas/IntentUrgency" },
					"expiresInMinutes": {
						"type": "integer",
						"minimum": 1,
						"description": "Minutes until intent expires (default: 1440 = 24 hours)"
					},
					"details": { "$ref": "#/components/schemas/TransferIntent" }
				}
			},
...
			"UpdateIntentStatusByIdRequest": {
				"type": "object",
				"required": ["id", "status"],
				"description": "Used with POST /api/intents/status (preferred). Includes intent ID in the body.",
				"properties": {
					"id": { "type": "string", "description": "Intent ID (e.g., int_1707048000_abc12345)" },
					"status": { "$ref": "#/components/schemas/IntentStatus" },
					"txHash": { "type": "string", "description": "Transaction hash" },
					"note": { "type": "string", "description": "Audit note" },
					"paymentSignatureHeader": {
						"type": "string",
						"description": "x402: base64-encoded payment signature header"
					},
					"paymentPayload": {
						"type": "object",
						"additionalProperties": true,
						"description": "x402: structured payment payload"
					},
					"settlementReceipt": {
						"type": "object",
						"additionalProperties": true,
						"description": "x402: settlement receipt from server"
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time",
						"description": "Override expiration timestamp"
					}
				}
			},
			"RegisterAgentRequest": {
				"type": "object",
				"required": ["trustChainId", "agentPublicKey", "authorizationSignature"],
				"properties": {
					"trustChainId": {
						"type": "string",
						"description": "Owner's wallet address (lowercased)"
					},
					"agentPublicKey": {
						"type": "string",
						"description": "0x-prefixed hex-encoded secp256k1 compressed public key"
					},
					"agentLabel": {
						"type": "string",
						"description": "Human-readable label (default: 'Unnamed Agent')"
					},
					"authorizationSignature": {
						"type": "string",
						"description": "EIP-191 personal_sign of: 'Authorize agent key for Ledger Agent Payments\\nKey: <pubkey>\\nLabel: <label>\\nIdentity: <trustchainId>'"
					}
				}
			},
...
			"TrustchainMember": {
				"type": "object",
				"required": ["id", "trustchainId", "memberPubkey", "role", "createdAt"],
				"properties": {
					"id": { "type": "string", "format": "uuid" },
					"trustchainId": { "type": "string", "description": "Owner's wallet address" },
					"memberPubkey": {
						"type": "string",
						"description": "Hex-encoded secp256k1 compressed public key"
					},
					"role": {
						"type": "string",
						"enum": ["agent_write_only", "full_access"],
						"default": "agent_write_only"
					},
					"label": { "type": "string", "nullable": true },
					"createdAt": { "type": "string", "format": "date-time" },
					"revokedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true,
						"description": "Set when agent is revoked; null if active"
					}
				}
			},
			"AgentResponse": {
				"type": "object",
				"required": ["success", "member"],
				"properties": {
					"success": { "type": "boolean", "const": true },
					"member": { "$ref": "#/components/schemas/TrustchainMember" }
				}
			},
...
		}
	}
}